requirements.txt
pillow
requests
pygame
pyttsx3